+++ service/parserprovider/default_test.go
@@ -25,7 +25,7 @@ import (
 	"github.com/stretchr/testify/require"
 
 	"go.opentelemetry.io/collector/config"
-	"go.opentelemetry.io/collector/config/configparser"
+	"go.opentelemetry.io/collector/config/configmap"
 	"go.opentelemetry.io/collector/config/configunmarshaler"
 	"go.opentelemetry.io/collector/processor/batchprocessor"
 	"go.opentelemetry.io/collector/receiver/otlpreceiver"
@@ -45,7 +45,7 @@ func TestDefault(t *testing.T) {
 		require.NoError(t, err)
 		pl := Default()
 		require.NotNil(t, pl)
-		var cp *configparser.ConfigMap
+		var cp *configmap.ConfigMap
 		cp, err = pl.Get(context.Background())
 		require.NoError(t, err)
 		require.NotNil(t, cp)
@@ -65,7 +65,7 @@ func TestDefault(t *testing.T) {
 		require.NoError(t, err)
 		pl := Default()
 		require.NotNil(t, pl)
-		var cp *configparser.ConfigMap
+		var cp *configmap.ConfigMap
 		cp, err = pl.Get(context.Background())
 		require.NoError(t, err)
 		require.NotNil(t, cp)
@@ -96,7 +96,7 @@ func TestDefault(t *testing.T) {
 		require.NoError(t, err)
 		pl := Default()
 		require.NotNil(t, pl)
-		var cp *configparser.ConfigMap
+		var cp *configmap.ConfigMap
 		cp, err = pl.Get(context.Background())
 		require.NoError(t, err)
 		require.NotNil(t, cp)